\name{Intview}
\docType{data}
\alias{Intview}
\title{Intview}
\description{
  a cross section from 2014

  \emph{number of observations} : 10351

  \emph{observation} : households

  \emph{country} : United-States
}
\usage{
data(Intview)
}
\format{A dataframe containing :
  \describe{

    \item{urban}{does the household live in an urban area (\code{yes} or \code{no}),}

    \item{smsa}{does the household live in a SMSA (\code{yes} or \code{no}),}

    \item{region}{the region of residence of the household, one of \code{northeast}, \code{midwest}, \code{south} and \code{west},}

    \item{age}{the age of the reference person of the household,}

    \item{educ}{the number of year of education of the reference person of the household,}

    \item{sex}{the sex of the reference person of the household (\code{male} and \code{female}),}

    \item{race}{the race of the head of the household, one of \code{white}, \code{black}, \code{indian}, \code{asian}, \code{pacific} and \code{multirace},}

    \item{hispanic}{is the reference person of the household is hispanic (\code{no} or \code{yes}),}

    \item{size}{the number of persons in the household,}

    \item{less18}{the number of persons less than 18 years old in the household,}

    \item{income}{the income of the household for the 12 month before the interview,}

    \item{month}{the month of the interview (between 1 and 12),}

    \item{foodtot}{food total,}
    \item{foodhome}{food at home,}
    \item{cereal}{cereal and cereal products,}
    \item{bakeprod}{backery products,}
    \item{beef}{beef,}
    \item{pork}{pork,}
    \item{othmeat}{other meats,}
    \item{poultry}{poultry,}
    \item{seafood}{fish and seafood,}
    \item{eggs}{eggs,}
    \item{milkproduct}{fresh milk and cream,}
    \item{othdairy}{other dairy products,}
    \item{frshfrut}{fresh fruits,}
    \item{frshveg}{fresh vegetables,}
    \item{procfrut}{processed fruits,}
    \item{procveg}{processed vegetables,}
    \item{sweets}{sugar and other sweets,}
    \item{nonalbev}{nonalcoholic beverages,}
    \item{oils}{fats and oils,}
    \item{miscfood}{miscellaneous foods,}
    \item{foodaway}{food away from home,}
    \item{alcbev}{alcoholic beverages,}
    \item{smoksupp}{tobacco products and smoking supplies,}
    \item{petfood}{pet food.}
}    
}
\source{ Consumer Expenditure Survey (CE), program of the US Bureau of
    Labor Statistics \url{http://www.bls.gov/cex/}, diary survey.
}

\keyword{datasets}
