% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvuongtest.R
\name{ndvuongtest}
\alias{ndvuongtest}
\title{Vuoung test for non-nested models}
\usage{
ndvuongtest(
  x,
  y,
  size = 0.05,
  pval = TRUE,
  type = c("non-nested", "nested", "overlapping"),
  ndraws = 10000,
  diffnorm = 0.1,
  seed = 1,
  print.level = 0
)
}
\arguments{
\item{x}{a first fitted model of class \code{"mhurdle"},}

\item{y}{a second fitted model of class \code{"mhurdle"},}

\item{size}{the size of the test,}

\item{pval}{should the p-value be computed ?}

\item{type}{the kind of test to be computed,}

\item{ndraws}{the number of draws for the simulations,}

\item{diffnorm}{a creuser,}

\item{seed}{the seed,}

\item{print.level}{the level of details to be printed.}
}
\value{
an object of class \code{"htest"}
}
\description{
The Vuong test is suitable to discriminate between two non-nested models.
}
\examples{

data("Interview", package = "mhurdle")
# dependent double hurdle model
dhm <- mhurdle(vacations ~ car + size | linc + linc2 | 0, Interview,
              dist = "ln", h2 = TRUE, method = "bhhh", corr = TRUE)

# a double hurdle p-tobit model
ptm <- mhurdle(vacations ~ 0 | linc + linc2 | car + size, Interview,
              dist = "ln", h2 = TRUE, method = "bhhh", corr = TRUE)
vuongtest(dhm, ptm)
}
\references{
Vuong Q.H. (1989) Likelihood ratio tests for model selection and
non-nested hypothesis, Econometrica, vol.57(2), pp.307-33.
}
\seealso{
\code{vuong} in package \code{pscl}.
}
\keyword{htest}
