% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_pkgs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{estfun}
\alias{bread}
\alias{meat}
\alias{llcont}
\alias{glance}
\alias{tidy}
\alias{prediction}
\alias{margins}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{margins}{\code{\link[margins]{margins}}}

  \item{nonnest2}{\code{\link[nonnest2]{llcont}}}

  \item{prediction}{\code{\link[prediction]{prediction}}}

  \item{sandwich}{\code{\link[sandwich]{bread}}, \code{\link[sandwich]{estfun}}, \code{\link[sandwich]{meat}}}
}}

