% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhurdle.methods.R
\name{rsq}
\alias{rsq}
\title{R squared and pseudo R squared}
\usage{
rsq(
  object,
  type = c("coefdet", "lratio"),
  adj = FALSE,
  r2pos = c("rss", "ess", "cor")
)
}
\arguments{
\item{object}{an object of class \code{"mhurdle"},}

\item{type}{one of \code{"coefdet"} or \code{"lratio"} to select a pseudo
coefficient of correlation or a Mc Fadden like measure based on the
likelihood function,}

\item{adj}{if \code{TRUE} a correction for the degrees of freedom is
performed,}

\item{r2pos}{only for pseudo coefficient of determination, should the
positive part of the R squared be computed using the residual sum of squares
(\code{"rss"}), the explained sum of squares (\code{"ess"}) or the
coefficient of correlation between the fitted values and the response
(\code{cor}).}
}
\value{
a numerical value
}
\description{
This function computes the R squared for multiple hurdle models. The measure
is a pseudo coefficient of determination or may be based on the likelihood.
}
\examples{

data("Interview", package = "mhurdle")
# independent double hurdle model
idhm <- mhurdle(vacations ~ car + size | linc + linc2 | 0, Interview,
              dist = "ln", h2 = TRUE, method = "bfgs")
rsq(idhm, type = "lratio")
rsq(idhm, type = "coefdet", r2pos = "rss")
}
\references{
McFadden D (1974). The Measurement of Urban Travel Demand. Journal of Public
Economics, 3, 303-328.
}
\keyword{htest}
