% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_pkgs.R
\name{sandwich_nonnest2}
\alias{sandwich_nonnest2}
\alias{estfun.mhurdle}
\alias{bread.mhurdle}
\alias{llcont.mhurdle}
\title{sandwich and nonnest2's methods}
\usage{
\method{estfun}{mhurdle}(x, ...)

\method{bread}{mhurdle}(x, ...)

\method{llcont}{mhurdle}(x, ...)
}
\arguments{
\item{x}{a model fitted with mhurdle}

\item{...}{further arguments, currently unused}
}
\description{
Methods to compute extract different features of the log-likelihood
function
}
\details{
\code{mhurdle} exports the \code{sandwich::estfun},
\code{sandwich::bread} and \code{nonnest2::llcont} functions. The
specific method provided for \code{mhurdle} objects are used in the
code to extract different features of the log-likelihood
}
